Change crate dependencies from attrSets to lists
Fixes issue #18

This allows for multiple dependencies with different targets. Before
this there was an issue with the flate2 crate which had dependencies:

    [dependencies]
    [...]
    miniz_oxide = { version = "0.3.2", optional = true}

    [target.'cfg(all(target_arch = "wasm32", not(target_os = "emscripten")))'.dependencies]
    miniz_oxide = "0.3.2"

but the generated Cargo.nix would only contain the targeted version:

    "miniz_oxide" = {
      packageId = "miniz_oxide 0.3.2 (registry+https://github.com/rust-lang/crates.io-index)";
      target = ((target."arch" == "wasm32") && (!(target."os" == "emscripten")));
    };

which means that miniz_oxide would not match for non-wasm32 targets,
even if it was enabled via a feature. Now the dependency appears twice:

    {
      name = "miniz_oxide";
      packageId = "miniz_oxide 0.3.3 (registry+https://github.com/rust-lang/crates.io-index)";
      optional = true;
    }
    {
      name = "miniz_oxide";
      packageId = "miniz_oxide 0.3.3 (registry+https://github.com/rust-lang/crates.io-index)";
      target = ((target."arch" == "x86_64"));
    }

so it will match when the target matches or when the feature is enabled.