Handle overlapping target.cfg sections better
Some packages (e.g. glutin) has the same package in multiple dependency
sections. We can handle this case by creating a combined target variable which
includes the dependency if the dependency matches any of the targets.

This doesn't handle the case where different features are enabled based
on the target, but it should be a strict improvement on the current
implementation.