Allow arbitrary commands before and after tests
An immediate use-case is to inject any environment variables that the
test may be relying on, such as location of some external files
or overriding some behaviour.

Another, perhaps more cliche use-case is for unsetting the `-e` bash
flag: this will allow us to capture output of failing tests without
giving up on expression itself. In this way, it fixes #191.