Fix applying overrides to defaultCrateOverrides when cross-compiling
When cross-compiling, there are different versions of defaultCrateOverrides for
each platform. This causes buildRustCrateWithFeatures to think the user has
passed a custom value of crateOverrides when cross-compiling, which prevents
defaultCrateOverrides from being applied. To fix this, use a null default for
crateOverrides to detect whether the user has actually specified a value.