improve git branch parsing and cargo config serialization
When it comes to git dependencies, cargo is very finicky, and it turns
out you basically have to copy the exact logic of 'cargo vendor' to get
it to work. This means:

  * The "precise" part of the URL, the part after the #, just gets
    ignored, for cargo-config purposes.
  * If either "branch" or "ref" are specified as ?key=value parts after
    the URL, these should be used as the "branch".
  * If neither is specified, the field should be omitted. Similarly, if
    the URL has no "?rev=" part, this should be omitted (NOT replaced
    by precise URL part).

I'm unsure what to do if both a tag and a branch are specified, since it
looks like cargo vendor treats branch and tag references as disjoint
things. Right now I just serialize both fields, but I believe this is
incorrect and one should take precedence.

Next, when fetching dependencies via fetchGit, we need to interpret the
URL data slightly differently:

  * If a precise part of the URL is given, that should supercede any
    "?rev=xyz" part of the URL.
  * If a tag or branch is provided we should use it as "ref". Otherwise
    I just set allRefs = true and hope for the best.