feat(tools): resolve workspace-inherited fields in manifests
When copying out workspace member crates from a larger workspace,
cargo will not be able to process the separated manifest if it still
refers to e.g. the workspace license or workspace dependencies, because
it can't find the workspace manifest.

We can't help it find the workspace manifest, because crates in a cargo
vendor directory aren't discovered recursively, and cargo doesn't resolve
symlinks before checking the ancestors to find the workspace manifest.