fix flake lock references pointing to wrong nixpkgs
There a couple of nix expressions read flake.lock to load nixpkgs:

- crate2nix/default.nix
- nix/nix-test-runner.nix

The problem is that `flakeLock.nodes.nixpkgs` does not refer to the
intended nixpkgs version: it refers to the cachix input's nixpkgs input
which is quite old. The version of nixpkgs locked at the flake root is
`flakeLock.nodes.nixpkgs_6`.

Lock file node names are named somewhat
arbitrarily. If the intention is to load the version of nixpkgs that is
locked in flake.nix then it is necessary to:

1. Read `flakeLock.root` to get the name of the root node. (The name is
   "root", but it's better not to assume.)
2. Read `flakeLock.nodes.${rootNodeName}.inputs.nixpkgs` to get the name of the
   lock file node of the intended nixpkgs version.
3. Read `flakeLock.nodes.${nixpkgsNodeName}`

nix/nix-test-runner.nix also gets the wrong version of nix-test-runner
for the same reason.

This change introduces a function, `flakeInput`, that applies the steps
above.

I wanted to be able to submit a PR with working tests, but the tests
were not working so this PR also includes some test fixes. While I was
at it I fixed clippy warnings.

There is another issue that is explained in a comment that I put in
nix/nix-test-runner.nix. I have a fix for that which I will submit in
another PR.