Use rust.platform.env, where set, in conditional compilation
Nixpkgs learnt about `target_env` in 
https://github.com/NixOS/nixpkgs/pull/531986. It'll be a little while 
before we can rely on that commit being present for everyone.

In the meantime, if it's unset, fall back to `gnu` as the least
harmful behaviour: this will not break any builds that weren't broken
before. Eventually maybe this can go away and `env` can just be
inherited like `arch` and `os`.

Fixes #402.